refactor(client): move clipboard handler into tokenPreview
